feature/cmake (#12)
* build library with cmake.
* make exceutable.
* change library name to just libvarcore.
* ignore generated files.
* added cunit as submodule
* added building of tools, test and examples.
* build everything works
add dependency to generator tools/varpp
* disable tests and examples for cunit
* silence compiler warnings
adopt repl to winsock
* fix compiling on linux
* No log output for variable of type TYPE_ACTION

+++ tools/varpp/utils.h
@@ -30,13 +30,14 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include <stdlib.h>
 #include <stdint.h>
 
 #ifndef __UTILS_H_
 #define __UTILS_H_
 
 void   split( char* s, char sep, char** svec, uint32_t len, uint32_t* n );
-char*  srepeat( char c, uint16_t len );
+char*  srepeat( char c, size_t len );
 char*  strtrim( char* s, char c );
 char*  skip_space( char * );
